[Chore] Move all the utils required by both `math.ts` & `numbers.ts`
This commit moves all the utility functions & their dependents required
by `math` to a shared file called `shared_utils.ts` from which both
`numbers.ts` & `math.ts` can use.
